only show the "Boot Options" entry when Fkey for "Kernel Option" is pressed
the difficulty here is to workaround the "Boot Options" from dirtying the
display of the panel below it. A solution would be to move the "Boot Options"
upper to leave some room between it and the panel.

also adding the "Kernel Option" choice for grub, since otherwise you can't
edit "Boot Options"

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/theme/mandriva-gfxboot-theme/trunk@252189 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94