Rework the way we handle dual-arch
The autodetection of 32/64 ISOs uses getcwd which currently always
return '/' in the COM32 module so this can't work. Change it to check
for hardcoded paths working for mandriva CDs. Run a linux64 image when
using 64 bit mode since otherwise I'm not sure how to get gfxboot to
pick the proper kernel :-/

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/theme/mandriva-gfxboot-theme/trunk@253214 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94