Fix(?) off by one error which led to display corruption in the lang menu
The changes to reduce the size of the status bar exhibited a bug in the
way the language menu is layed out (vertical stripes all over the
place). panel.text.y was 0 and we tried to substract 1 from it when
trying to draw the lang menu background which didn't work too nicely.
This commit is one way of fixing it (by changing the way we calculate
the number of lang entries to display in one column), but I'm not sure
this is the best way to do it.

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/theme/mandriva-gfxboot-theme/trunk@253227 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94